+++ demo/latest/src/App.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { Worker } from '@phuocng/react-pdf-viewer';
 
 import '@phuocng/react-pdf-viewer/cjs/react-pdf-viewer.css';
-import KeywordExample from './KeywordExample';
+import WaterMarkExample from './WaterMarkExample';
 
 const App = () => {
     return (
         <Worker workerUrl="https://unpkg.com/pdfjs-dist@2.2.228/build/pdf.worker.min.js">
-            <KeywordExample />
+            <WaterMarkExample />
         </Worker>
     );
 };